# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NetVent Parivaar API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/members:
    get:
      operationId: listMembers
      tags: [members]
      summary: Search and list members
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: expertise
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectoryListResponse"
  /v1/members/{memberId}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get member profile
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectoryMemberResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # EVENTS
  /v1/events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [UPCOMING, PAST]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Events list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
  /v1/events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event detail
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/groups/{groupId}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get group detail
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetail"
...
  /v1/admin/members:
    get:
      operationId: adminListMembers
      tags: [admin]
      summary: Admin list members with filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMemberListResponse"
  /v1/admin/members/{profileId}/approve:
    post:
      operationId: adminApproveMember
      tags: [admin]
      summary: Approve member application
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberProfile"
  /v1/admin/members/{profileId}/reject:
    post:
      operationId: adminRejectMember
      tags: [admin]
      summary: Reject member application
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberProfile"
  /v1/admin/members/{profileId}:
    patch:
      operationId: adminUpdateMember
      tags: [admin]
      summary: Admin update member profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminMemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberProfile"
    delete:
      operationId: adminDeleteMember
      tags: [admin]
      summary: Delete member
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /v1/admin/events:
    post:
      operationId: adminCreateEvent
      tags: [admin]
      summary: Create event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /v1/admin/events/{eventId}:
    patch:
      operationId: adminUpdateEvent
      tags: [admin]
      summary: Update event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: adminDeleteEvent
      tags: [admin]
      summary: Delete event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /v1/admin/groups:
    post:
      operationId: adminCreateGroup
      tags: [admin]
      summary: Create group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /v1/admin/groups/{groupId}:
    patch:
      operationId: adminUpdateGroup
      tags: [admin]
      summary: Update group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
    delete:
      operationId: adminDeleteGroup
      tags: [admin]
      summary: Delete group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /v1/admin/team:
    post:
      operationId: adminCreateTeamMember
      tags: [admin]
      summary: Add team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /v1/admin/team/{teamMemberId}:
    patch:
      operationId: adminUpdateTeamMember
      tags: [admin]
      summary: Update team member
      parameters:
        - name: teamMemberId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
    delete:
      operationId: adminDeleteTeamMember
      tags: [admin]
      summary: Delete team member
      parameters:
        - name: teamMemberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /v1/admin/cities:
    post:
      operationId: adminCreateCity
      tags: [admin]
      summary: Create city
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
  /v1/admin/cities/{cityId}:
    patch:
      operationId: adminUpdateCity
      tags: [admin]
      summary: Update city
      parameters:
        - name: cityId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
  /v1/admin/reports:
    get:
      operationId: adminListReports
      tags: [admin]
      summary: List reports
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [OPEN, RESOLVED]
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportListResponse"
  /v1/admin/reports/{reportId}/resolve:
    patch:
      operationId: adminResolveReport
      tags: [admin]
      summary: Resolve report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /v1/reports:
    post:
      operationId: createReport
      tags: [members]
      summary: Report a post or profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Reported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [success, error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # AUTH SCHEMAS
...
    DirectoryListResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/DirectoryMember"
        pagination:
          $ref: "#/components/schemas/Pagination"
    DirectoryMemberResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/DirectoryMember"

    # MEMBER SCHEMAS
...
    MemberProfile:
      type: object
      required: [id, userId, status]
      properties:
        id:
          type: string
        userId:
          type: string
        memberId:
          type: ["string", "null"]
        fullName:
          type: string
        profilePhotoUrl:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        cityId:
          type: ["string", "null"]
        stateId:
          type: ["string", "null"]
        institution:
          type: ["string", "null"]
        classYear:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        portfolioUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        rejectionReason:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        privacySettings:
          $ref: "#/components/schemas/PrivacySettings"
        skills:
          type: array
          items:
            $ref: "#/components/schemas/Skill"
        city:
          $ref: "#/components/schemas/City"
        state:
          $ref: "#/components/schemas/State"
        createdAt:
          type: string
...
    AdminMemberListResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/MemberProfile"
        pagination:
          $ref: "#/components/schemas/Pagination"
    AdminMemberUpdate:
      type: object
      properties:
        fullName:
          type: string
        role:
          type: string
          enum: [BASIC, MEMBER, MODERATOR, ADMIN]
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
    RejectInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string

    # EVENT SCHEMAS
    Event:
      type: object
      required: [id, title, eventDate, status]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        venue:
          type: ["string", "null"]
        eventDate:
          type: string
        status:
          type: string
          enum: [UPCOMING, PAST]
        coverImageUrl:
          type: ["string", "null"]
        rsvpCount:
          type: integer
        hasRsvped:
          type: boolean
        createdAt:
          type: string
    EventDetail:
      type: object
      required: [id, title, eventDate, status]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        venue:
          type: ["string", "null"]
        eventDate:
          type: string
        status:
          type: string
          enum: [UPCOMING, PAST]
        coverImageUrl:
          type: ["string", "null"]
        rsvpCount:
          type: integer
        hasRsvped:
          type: boolean
        photos:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              photoUrl:
                type: string
            required: [id, photoUrl]
        sponsors:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              logoUrl:
                type: ["string", "null"]
            required: [id, name]
        createdAt:
          type: string
    EventInput:
      type: object
      required: [title, eventDate, status]
      properties:
        title:
          type: string
        description:
          type: string
        venue:
          type: string
        eventDate:
          type: string
        status:
          type: string
          enum: [UPCOMING, PAST]
        coverImageUrl:
          type: ["string", "null"]
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        venue:
          type: string
        eventDate:
          type: string
        status:
          type: string
          enum: [UPCOMING, PAST]
        coverImageUrl:
          type: ["string", "null"]
    EventListResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # GROUP SCHEMAS
    Group:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        memberCount:
          type: integer
        postCount:
          type: integer
        isMember:
          type: boolean
        createdAt:
          type: string
    GroupDetail:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        memberCount:
          type: integer
        isMember:
          type: boolean
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    GroupInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        moderatorId:
          type: ["string", "null"]
    GroupUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        moderatorId:
          type: ["string", "null"]
...
    TeamMember:
      type: object
      required: [id, name, roleTitle, section, displayOrder]
      properties:
        id:
          type: string
        name:
          type: string
        roleTitle:
          type: string
        photoUrl:
          type: ["string", "null"]
        section:
          type: string
          enum: [FOUNDER, CORE, ADVISOR]
        displayOrder:
          type: integer
        socialLinks:
          type: object
          properties:
            instagram:
              type: ["string", "null"]
            linkedin:
              type: ["string", "null"]
            twitter:
              type: ["string", "null"]
    TeamMemberInput:
      type: object
      required: [name, roleTitle, section, displayOrder]
      properties:
        name:
          type: string
        roleTitle:
          type: string
        photoUrl:
          type: ["string", "null"]
        section:
          type: string
          enum: [FOUNDER, CORE, ADVISOR]
        displayOrder:
          type: integer
        socialLinks:
          type: object
          properties:
            instagram:
              type: ["string", "null"]
            linkedin:
              type: ["string", "null"]
            twitter:
              type: ["string", "null"]
    TeamMemberUpdate:
      type: object
      properties:
        name:
          type: string
        roleTitle:
          type: string
        photoUrl:
          type: ["string", "null"]
        section:
          type: string
          enum: [FOUNDER, CORE, ADVISOR]
        displayOrder:
          type: integer
        socialLinks:
          type: object
          properties:
            instagram:
              type: ["string", "null"]
            linkedin:
              type: ["string", "null"]
            twitter:
              type: ["string", "null"]
...
    City:
      type: object
      required: [id, name, status]
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [ACTIVE, UPCOMING]
        stateId:
          type: string
        state:
          $ref: "#/components/schemas/State"
        memberCount:
          type: integer
...
    CityInput:
      type: object
      required: [name, stateId, status]
      properties:
        name:
          type: string
        stateId:
          type: string
        status:
          type: string
          enum: [ACTIVE, UPCOMING]
    CityUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [ACTIVE, UPCOMING]
        cityLeadId:
          type: ["string", "null"]

    # SKILLS
...
    ReportInput:
      type: object
      required: [targetType, targetId, reason]
      properties:
        targetType:
          type: string
          enum: [POST, PROFILE]
        targetId:
          type: string
        reason:
          type: string
    ReportListResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Report"

    # STATS
...